valet/api/common/compute.py
# -*- encoding: utf-8 -*-
#
# Copyright (c) 2014-2016 AT&T
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#    http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
# implied.
#
# See the License for the specific language governing permissions and
# limitations under the License.

'''Compute helper library'''

from novaclient import client
from pecan import conf

# Nova API v2
VERSION = 2


def nova_client():
    '''Returns a nova client'''
    sess = conf.identity.engine.session
    nova = client.Client(VERSION, session=sess)
    return nova
