test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.


hacking<0.11,>=0.10.0
os-testr<=0.7.0
markupsafe<=0.23
pecan<=0.8.2
notario<=0.0.11
coverage>=3.6
python-subunit>=0.0.18
mock>=1.2
oslotest>=1.10.0 # Apache-2.0
oslo.config>=1.9.0
testrepository>=0.0.18
sphinx!=1.2.0,!=1.3b1,<1.3,>=1.1.2
testscenarios>=0.4
testtools>=1.4.0
oslo.i18n<=3.8.0
oslo.log>=1.0.0
pytz
python-keystoneclient<=3.4.0
python-novaclient<=4.0.0
python-heatclient<=1.2.0

oslo.messaging==1.8.3
#tempest<=12.1.0 ---------- needs to be installed on Jenkins, no output when using tox
#tempest-lib>=0.8.0